improv(appearance): move fonts into context view
